Refine Progress Property in Notifications for Enhanced Precision
- Updated the progress property from int to float in the notification system for more precise progress bar visualization.
- Modified NotificationBuilder to accommodate and correctly handle the progress property as a float value.
- This change results in a more accurate and granular representation of progress within notifications, significantly improving the user interface and experience.

+++ core/src/main/java/me/carda/awesome_notifications/core/builders/NotificationBuilder.java
@@ -1637,7 +1637,7 @@ public class NotificationBuilder {
         if (!stringUtils.isNullOrEmpty(contentModel.summary))
             builder.setSubText(contentModel.summary);
 
-        if(contentModel.progress != null && IntegerUtils.isBetween(contentModel.progress, 0, 100))
+        if(contentModel.progress != null && IntegerUtils.isBetween(contentModel.progress.intValue(), 0, 100))
             builder.setProgress(
                 100,
                 Math.max(0, Math.min(100, IntegerUtils.extractInteger(contentModel.progress, 0))),
